Users can now create recording directories, delete recording directories and rename their recording directories.
whenever a change to recording directories occur, the new configuration is pushed to NextPVR automatically

+++ NextPvrWebConsole/Scripts/usersettings/recordingdirectories.js
@@ -12,18 +12,44 @@ $(function () {
         self.selectedChannelGroup = ko.observable();
 
         self.delete = function (directory) {
-            api.deleteJSON("recordingdirectories/" + directory.oid(), null, function () {
-                self.recordingdirectories.remove(directory);
+            gui.confirmMessage({
+                message: $.i18n._("Are you sure you want to delete the Recording Directory '%s'?", [directory.name()]),
+                yes: function () {
+                    api.deleteJSON("recordingdirectories/" + directory.oid(), null, function () {
+                        self.recordingdirectories.remove(directory);
+                    });
+                }
+            });
+        };
+
+        self.edit = function (directory) {
+            gui.promptMessage({
+                title: $.i18n._("Edit Recording Folder"),
+                message: $.i18n._("Type in the name of the folder."),
+                validationMessage: DIRECTORY_ERROR_MESSAGE,
+                validationExpression: DIRECTORY_REGULAR_EXPRSESION,
+                initialValue: directory.name(),
+                success: function (name) {
+                    api.putJSON('recordingdirectories/' + directory.oid(), name, function () {
+                        directory.name(name);
+                    });
+                }
             });
         };
 
         self.add = function () {
-            var directory = new RecordingDirectory();
-            self.recordingdirectories.push(directory);
-        };
-
-        self.select = function (channelGroup) {
-            
+            gui.promptMessage({
+                title: DIRECTORY_CREATE_TITLE,
+                message: DIRECTORY_CREATE_MESSAGE,
+                validationMessage: DIRECTORY_ERROR_MESSAGE,
+                validationExpression: DIRECTORY_REGULAR_EXPRSESION,
+                success: function (name) {
+                    api.postJSON('recordingdirectories', name, function (item) {
+                        var directory = new RecordingDirectory(item);
+                        self.recordingdirectories.push(directory);
+                    });
+                }
+            });
         };
 
         var refreshRecordingDirectories = function () {
@@ -35,7 +61,9 @@ $(function () {
         refreshRecordingDirectories();
     }
 
-    var div = $('.user-settings .vtab-content > .recording-directories');
-    if(div.length > 0)
+    var div = $('#user-settings-tab-recordingdirectories');
+    if (div.length > 0) {
         ko.applyBindings(new RecordingDirectoriesViewModel(), div.get(0));
+        div.find('ul').removeAttr('style');
+    }
 });
